requirements.txt
# ==============================================================================
#  Zanao Project - Python Dependencies (Optimized Version)
# ==============================================================================
#
#  使用以下命令安装所有依赖:
#  pip install -r requirements.txt
#
# ------------------------------------------------------------------------------
#  Part 1: 微信聊天记录导出工具 (wx_login & group_chat_getter)
# ------------------------------------------------------------------------------
#  用于解密和操作微信数据库
sqlcipher3-wheels==0.5.5
pycryptodome==3.23.0
pywxdump==3.1.45
#  [补充] Windows平台API交互，用于读取内存和注册表，是wx_login的关键
pywin32==311

# ------------------------------------------------------------------------------
#  Part 2: 集市爬虫 (zanao_climber)
# ------------------------------------------------------------------------------
#  用于与Redis队列交互
redis==6.2.0
#  用于发送HTTP网络请求
requests==2.32.4
#  用于显示美观的进度条
tqdm==4.67.1

# ------------------------------------------------------------------------------
#  Part 3: 帖子分析器 (zanao_analyzer) - 核心AI与数据处理
# ------------------------------------------------------------------------------
#  深度学习框架 (NER、情感、相似度模型的基础)
torch==2.7.1
#  Hugging Face模型库，用于加载情感分析等模型
transformers==4.43.4
#  用于文本向量化和相似度计算
sentence-transformers==5.0.0
#  [补充] sentence-transformers 的核心依赖，用于科学计算和机器学习
scikit-learn==1.7.1
#  命名实体识别(NER)模型库
gliner==0.2.21
#  强大的数据分析与处理库
pandas==2.3.1
#  中文分词库，用于词云和新词发现
jieba==0.42.1
#  科学计算基础库，Pandas和Torch等库的依赖
numpy==1.26.4
#  用于本地运行大语言模型 (LLM) 服务
ollama==0.5.1

# ------------------------------------------------------------------------------
#  Part 4: 应用层与API服务器 (applications & api_server)
# ------------------------------------------------------------------------------
#  高性能的Web框架，用于构建API
fastapi==0.116.1
#  FastAPI官方推荐的ASGI服务器，用于运行应用
uvicorn==0.35.0
#  数据校验与模型定义，FastAPI的核心依赖
pydantic==2.11.7
#  功能强大的图表生成库
pyecharts==2.0.8
#  生成词云图
wordcloud==1.9.4

# ------------------------------------------------------------------------------
#  Part 5: 配置与实用工具
# ------------------------------------------------------------------------------
#  [推荐] 用于从.env文件加载环境变量，是管理配置的最佳实践
python-dotenv==1.1.1